Add support for EventSource endpoints
This commit demonstrates the ability to easily add EventSource
endpoints using SvelteKit Endpoints. By having an endpoint return an
async generator function as the body, SvelteKit will keep emitting
writes for as long as the endpoint yields values. Or, for as long as
the client holds the request open; which ever ends first.

A couple of notes, async generator functions are instantiated as async
genators inside SvelteKit, so there are no arguments passed in to the
function; it must rely on scoping. In addition, async is required, as
a synchronous function would be detrimental due to a hard for loop
inside the request processor/adapter.